RosettaFunction/requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.33.0
azure-functions==1.23.0
certifi==2025.7.14
charset-normalizer==3.4.2
dotenv==0.9.9
idna==3.10
isodate==0.7.2
MarkupSafe==2.1.5
msrest==0.7.1
oauthlib==3.3.1
python-dotenv==1.0.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==3.0.6
